pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pedro"
version = "0.1.0"
description = "Plan-Extended Deep Research Operator: a multi-agent deep research system that scouts the field with you before committing to a plan."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "Jordan Lim" }]
dependencies = [
    "openai>=1.55.0",
    "pydantic>=2.7",
    "langgraph>=0.2.50",
    "fastapi>=0.110",
    "sse-starlette>=2.1",
    "uvicorn[standard]>=0.30",
    "python-dotenv>=1.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.6",
]

[project.scripts]
pedro = "pedro.api.app:main"

[tool.hatch.build.targets.wheel]
packages = ["pedro"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra"
markers = [
    "live: marks tests that hit the real OpenAI API (deselected by default; gated by PEDRO_RUN_LIVE=1)",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
ignore = ["E501"]
